+++ src/benchmarks/edu/brown/benchmark/voter/voter-ddl.sql
@@ -25,7 +25,7 @@ CREATE TABLE area_code_state
 --   voters are not allowed to submit more than <x> votes, x is passed to client application
 CREATE TABLE votes
 (
-  vote_id			 integer    NOT NULL,
+  vote_id            bigint     NOT NULL,
   phone_number       bigint     NOT NULL
 , state              varchar(2) NOT NULL -- REFERENCES area_code_state (state)
 , contestant_number  integer    NOT NULL REFERENCES contestants (contestant_number)
